requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
Pillow==12.1.1
openpyxl==3.1.5
duckduckgo-search==8.1.1
